Decouple *-Description from ManagementClient by moving entity name valication code into EntityNameHelper where it should be.
Left those helper methods internally, but could probably make it public to allow customers to use these validation rules rather than re-implement their own.
